+++ src/startData.js
@@ -1,10 +1,9 @@
-// startData.js
 import { fetchAllUserData, fetchSingleUserData, fetchAllTripsData, fetchAllDestinationData } from './apiCalls';
 
 let allUsersData = [];
 let allTripData = [];
 let allDestinationData = [];
-let allSingleUserData = [];
+let allSingleUserData = {};
 
 export const fetchAllData = (userId) => {
   return Promise.all([
@@ -13,10 +12,10 @@ export const fetchAllData = (userId) => {
     fetchAllTripsData(),
     fetchAllDestinationData()
   ]).then(([allUserDataResult, singleUserDataResult, allTripsDataResult, allDestinationDataResult]) => {
-    allUsersData = allUserDataResult;
+    allUsersData = allUserDataResult.travelers;
     allSingleUserData = singleUserDataResult;
-    allTripData = allTripsDataResult;
-    allDestinationData = allDestinationDataResult;
+    allTripData = allTripsDataResult.trips;
+    allDestinationData = allDestinationDataResult.destinations;
     console.log('Fetched all data:', { allUsersData, allSingleUserData, allTripData, allDestinationData });
   }).catch(error => {
     console.error('Error fetching data:', error);
